+++ functions/go/apply-setters/applysetters/visit.go
@@ -1,0 +1,358 @@
+package applysetters
+
+import (
+	"fmt"
+	"regexp"
+	"strings"
+
+	"github.com/kptdev/kpt/api/fnresult/v1"
+	"sigs.k8s.io/kustomize/kyaml/errors"
+	"sigs.k8s.io/kustomize/kyaml/fn/framework"
+	"sigs.k8s.io/kustomize/kyaml/yaml"
+)
+
+// visitor is implemented by structs which need to walk the configuration.
+// visitor is provided to accept to walk configuration
+type visitor interface {
+	// visitScalar is called for each scalar field value on a resource
+	// node is the scalar field value
+	// path is the path to the field; path elements are separated by '.'
+	visitScalar(node *yaml.RNode, path string) error
+
+	// visitMapping is called for each Mapping field value on a resource
+	// node is the mapping field value
+	// path is the path to the field
+	visitMapping(node *yaml.RNode, path string) error
+}
+
+var _ visitor = &ApplySetters{}
+
+// VisitElements calls fn for each element in a SequenceNode.
+// Returns an error for non-SequenceNodes
+func VisitElements(rn *yaml.RNode, fn func(node *yaml.RNode, i int) error) error {
+	elements, err := rn.Elements()
+	if err != nil {
+		return errors.Wrap(err)
+	}
+	for i := range elements {
+		if err := fn(elements[i], i); err != nil {
+			return errors.Wrap(err)
+		}
+	}
+	return nil
+}
+
+/*
+visitMapping takes input mapping node, and performs following steps
+checks if the key node of the input mapping node has line comment with SetterCommentIdentifier
+checks if the value node is of sequence node type
+if yes to both, resolves the setter value for the setter name in the line comment
+replaces the existing sequence node with the new values provided by user
+
+e.g. for input of Mapping node
+
+environments: # kpt-set: ${env}
+- dev
+- stage
+
+For input ApplySetters [name: env, value: "[stage, prod]"], the yaml node is transformed to
+
+environments: # kpt-set: ${env}
+- stage
+- prod
+*/
+func (as *ApplySetters) visitMapping(object *yaml.RNode, path string) error {
+	return object.VisitFields(func(node *yaml.MapNode) error {
+		if node == nil || node.Key.IsNil() || node.Value.IsNil() {
+			// don't do IsNilOrEmpty check as empty sequences are allowed
+			return nil
+		}
+
+		// the aim of this method is to apply-setter for sequence nodes
+		if node.Value.YNode().Kind != yaml.SequenceNode {
+			// return if it is not a sequence node
+			return nil
+		}
+
+		lineComment := node.Key.YNode().LineComment
+		if node.Value.YNode().Style == yaml.FlowStyle {
+			// if node is FlowStyle e.g. env: [foo, bar] # kpt-set: ${env}
+			// the setter comment will be on value node
+			lineComment = node.Value.YNode().LineComment
+		}
+
+		origValue := node.Value.YNode().Value
+
+		setterPattern := extractSetterPattern(lineComment)
+		if setterPattern == "" {
+			// the node is not tagged with setter pattern
+			return nil
+		}
+
+		if !shouldSet(setterPattern, as.Setters) {
+			// this means there is no intent from user to modify this setter tagged resources
+			return nil
+		}
+
+		// since this setter pattern is found on sequence node, make sure that it is
+		// not interpolation of setters, it should be simple setter e.g. ${environments}
+		if !validArraySetterPattern(setterPattern) {
+			return errors.Errorf("invalid setter pattern for array node: %q", setterPattern)
+		}
+
+		// get the setter value for the setter name in the comment
+		sv := setterValue(as.Setters, setterPattern)
+
+		// add the key to the field path
+		fieldPath := strings.TrimPrefix(fmt.Sprintf("%s.%s", path, node.Key.YNode().Value), ".")
+
+		if sv == "" {
+			node.Value.YNode().Content = []*yaml.Node{}
+			// empty sequence must be FlowStyle e.g. env: [] # kpt-set: ${env}
+			node.Value.YNode().Style = yaml.FlowStyle
+			// setter pattern comment must be on value node
+			node.Value.YNode().LineComment = lineComment
+			node.Key.YNode().LineComment = ""
+			as.Results = append(as.Results, &v1.ResultItem{
+				Message:     fmt.Sprintf("set field value to %q", sv),
+				Severity:    framework.Info,
+				File:        &framework.File{Path: as.filePath},
+				ResourceRef: as.metadata,
+				Field: &v1.Field{
+					Path:          fieldPath,
+					CurrentValue:  origValue,
+					ProposedValue: sv,
+				},
+			})
+			return nil
+		}
+
+		// parse the setter value as yaml node
+		rn, err := yaml.Parse(sv)
+		if err != nil {
+			return errors.Errorf("input to array setter must be an array of values, but found %q", sv)
+		}
+
+		// the setter value must parse as sequence node
+		if rn.YNode().Kind != yaml.SequenceNode {
+			return errors.Errorf("input to array setter must be an array of values, but found %q", sv)
+		}
+
+		node.Value.YNode().Content = rn.YNode().Content
+		node.Key.YNode().LineComment = lineComment
+		// non-empty sequences should be standardized to FoldedStyle
+		// env: # kpt-set: ${env}
+		//  - foo
+		//  - bar
+		node.Value.YNode().Style = yaml.FoldedStyle
+
+		as.Results = append(as.Results, &v1.ResultItem{
+			Message:     fmt.Sprintf("set field value to %q", sv),
+			Severity:    framework.Info,
+			File:        &framework.File{Path: as.filePath},
+			ResourceRef: as.metadata,
+			Field: &v1.Field{
+				Path:          fieldPath,
+				CurrentValue:  origValue,
+				ProposedValue: sv,
+			},
+		})
+		return nil
+	})
+}
+
+/*
+visitScalar accepts the input scalar node and performs following steps,
+checks if the line comment of input scalar node has prefix SetterCommentIdentifier
+resolves the setter values for the setter name in the comment
+replaces the existing value of the scalar node with the new value
+
+e.g.for input of scalar node 'nginx:1.7.1 # kpt-set: ${image}:${tag}' in the yaml node
+
+apiVersion: v1
+...
+
+	image: nginx:1.7.1 # kpt-set: ${image}:${tag}
+
+and for input ApplySetters [[name: image, value: ubuntu], [name: tag, value: 1.8.0]]
+The yaml node is transformed to
+
+apiVersion: v1
+...
+
+	image: ubuntu:1.8.0 # kpt-set: ${image}:${tag}
+*/
+func (as *ApplySetters) visitScalar(object *yaml.RNode, path string) error {
+	if object.IsNil() {
+		return nil
+	}
+
+	if object.YNode().Kind != yaml.ScalarNode {
+		// return if it is not a scalar node
+		return nil
+	}
+
+	origValue := object.YNode().Value
+
+	// perform a direct set of the field if it matches
+	setterPattern := extractSetterPattern(object.YNode().LineComment)
+	if setterPattern == "" {
+		// the node is not tagged with setter pattern
+		return nil
+	}
+
+	curPattern := setterPattern
+	if !shouldSet(setterPattern, as.Setters) {
+		// this means there is no intent from user to modify this setter tagged resources
+		return nil
+	}
+
+	// replace the setter names in comment pattern with provided values
+	for _, setter := range as.Setters {
+		setterPattern = strings.ReplaceAll(
+			setterPattern,
+			fmt.Sprintf("${%s}", setter.Name),
+			fmt.Sprintf("%v", setter.Value),
+		)
+	}
+
+	// replace the remaining setter names in comment pattern with values derived from current
+	// field value, these values are not provided by user
+	currentValues := currentSetterValues(curPattern, object.YNode().Value)
+	for name, value := range currentValues {
+		setterPattern = strings.ReplaceAll(
+			setterPattern,
+			fmt.Sprintf("${%s}", name),
+			fmt.Sprintf("%v", value),
+		)
+	}
+
+	// check if there are unresolved setters and throw error
+	urs := unresolvedSetters(setterPattern)
+	if len(urs) > 0 {
+		return errors.Errorf("values for setters %v must be provided", urs)
+	}
+
+	object.YNode().Value = setterPattern
+	if setterPattern == "" {
+		object.YNode().Style = yaml.DoubleQuotedStyle
+	}
+	object.YNode().Tag = yaml.NodeTagEmpty
+	as.Results = append(as.Results, &v1.ResultItem{
+		Message:     fmt.Sprintf("set field value to %q", object.YNode().Value),
+		Severity:    framework.Info,
+		File:        &framework.File{Path: as.filePath},
+		ResourceRef: as.metadata,
+		Field: &v1.Field{
+			Path:          strings.TrimPrefix(path, "."),
+			CurrentValue:  origValue,
+			ProposedValue: object.YNode().Value,
+		},
+	})
+	return nil
+}
+
+// shouldSet takes the setter pattern comment and setter values map and returns true
+// iff at least one of the setter names in the pattern match with the setter names
+// in input setterValues map
+func shouldSet(pattern string, setters []Setter) bool {
+	for _, s := range setters {
+		if strings.Contains(pattern, fmt.Sprintf("${%s}", s.Name)) {
+			return true
+		}
+	}
+	return false
+}
+
+// currentSetterValues takes pattern and value and returns setter names to values
+// derived using pattern matching
+// e.g. pattern = my-app-layer.${stage}.${domain}.${tld}, value = my-app-layer.dev.example.com
+// returns {"stage":"dev", "domain":"example", "tld":"com"}
+func currentSetterValues(pattern, value string) map[string]string {
+	res := make(map[string]string)
+	// get all setter names enclosed in ${}
+	// e.g. value: my-app-layer.dev.example.com
+	// pattern: my-app-layer.${stage}.${domain}.${tld}
+	// urs: [${stage}, ${domain}, ${tld}]
+	urs := unresolvedSetters(pattern)
+	// and escape pattern
+	pattern = regexp.QuoteMeta(pattern)
+	// escaped pattern: my-app-layer\.\$\{stage\}\.\$\{domain\}\.\$\{tld\}
+
+	for _, setterName := range urs {
+		// escape setter name
+		// we need to escape the setterName as well to replace it in the escaped pattern string later
+		setterName = regexp.QuoteMeta(setterName)
+		pattern = strings.ReplaceAll(
+			pattern,
+			setterName,
+			`(?P<x>.*)`) // x is just a place holder, it could be any alphanumeric string
+	}
+	// pattern: my-app-layer\.(?P<x>.*)\.(?P<x>.*)\.(?P<x>.*)
+	r, err := regexp.Compile(pattern)
+	if err != nil {
+		// just return empty map if values can't be derived from pattern
+		return res
+	}
+	setterValues := r.FindStringSubmatch(value)
+	if len(setterValues) == 0 {
+		return res
+	}
+	// setterValues: [ "my-app-layer.dev.example.com", "dev", "example", "com"]
+	setterValues = setterValues[1:]
+	// setterValues: [ "dev", "example", "com"]
+	if len(urs) != len(setterValues) {
+		// just return empty map if values can't be derived
+		return res
+	}
+	for i := range setterValues {
+		if setterValues[i] == "" {
+			// if any of the value is unresolved return empty map
+			// and expect users to provide all values
+			return make(map[string]string)
+		}
+		res[clean(urs[i])] = setterValues[i]
+	}
+	return res
+}
+
+// setterValue returns the value for the setter
+func setterValue(setters []Setter, setterName string) string {
+	for _, setter := range setters {
+		if setter.Name == clean(setterName) {
+			return setter.Value
+		}
+	}
+	return ""
+}
+
+// extractSetterPattern extracts the setter pattern from the line comment of the
+// yaml RNode. If the the line comment doesn't contain SetterCommentIdentifier
+// prefix, then it returns empty string
+func extractSetterPattern(lineComment string) string {
+	if !strings.HasPrefix(lineComment, SetterCommentIdentifier) {
+		return ""
+	}
+	return strings.TrimSpace(strings.TrimPrefix(lineComment, SetterCommentIdentifier))
+}
+
+// validArraySetterPattern returns true if the array setter pattern is valid
+// pattern must not interpolation of setters, it should be simple setter e.g. ${environments}
+func validArraySetterPattern(pattern string) bool {
+	return len(unresolvedSetters(pattern)) == 1 &&
+		strings.HasPrefix(pattern, "${") &&
+		strings.HasSuffix(pattern, "}")
+}
+
+// unresolvedSetters returns the list of values enclosed in ${} present within given
+// pattern e.g. pattern = foo-${image}:${tag}-bar return ["${image}", "${tag}"]
+func unresolvedSetters(pattern string) []string {
+	re := regexp.MustCompile(`\$\{([^}]*)\}`)
+	return re.FindAllString(pattern, -1)
+}
+
+// clean extracts value enclosed in ${}
+func clean(input string) string {
+	input = strings.TrimSpace(input)
+	return strings.TrimSuffix(strings.TrimPrefix(input, "${"), "}")
+}
